% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Standard RNA Codon Table}
\format{
A named character vector with 64 elements.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table
}
\description{
A named character vector mapping RNA codons (triplets) to their
corresponding amino acid one-letter codes.
}
\keyword{datasets}
